Downloads/nchlh nkamlo/pfe-main/requirements.txt
django==3.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
mysqlclient==2.2.0